refactor(html-generation): move HTML generation logic to dedicated utility classes
Move HTML generation logic for error, helper, and status views to separate utility classes for better maintainability and code organization. This includes creating `GenerateHelperHtml`, `GenerateStatusHtml`, and relocating `generateErrorHtml` within `BranchHtmlGenerator`. The changes reduce code duplication and improve modularity.
